LPS-93391 Correcting usage of MBDiscussionGroupServiceConfiguration
