+++ modules/apps/configuration-admin/configuration-admin-web/src/main/resources/content/Language.properties
@@ -15,6 +15,7 @@ category.blogs=Blogs
 category.breadcrumbs=Breadcrumbs
 category.captcha=CAPTCHA
 category.chat=Chat
+category.comments=Comments
 category.community-tools=Community Tools
 category.connectors=Connectors
 category.devices=Devices
