+++ modules/apps/message-boards/message-boards-api/src/main/resources/com/liferay/message/boards/dependencies/discussion_email_subject.tmpl
